accessory_services/notification_service/requirements.txt
fastapi
uvicorn
requests
dapr-ext-fastapi-dev
